Added files via upload
When we received MiSeq data from the sequencing center in February 2016, we realized that the raw data still had primer sequences. This slide set describes how we used Mothur to remove the primer sequences. I found most of the information here in a Mothur forum post and the Mothur wiki (links in the slide notes).
Please see the slide notes for details.